Run tentacle through exe

+++ Tentacle/Scripts/run-octopusdeploy-tentacle.ps1
@@ -1,58 +1,23 @@
 ﻿[CmdletBinding()]
 Param()
 
+. ./octopus-common.ps1
 
-$OFS = "`r`n"
 
-function Write-Log
+function Run-OctopusDeployTentacle
 {
-  param (
-    [string] $message
-  )
-
-  $timestamp = ([System.DateTime]::UTCNow).ToString("yyyy'-'MM'-'dd'T'HH':'mm':'ss")
-  Write-Verbose "[$timestamp] $message"
-}
-
-function Execute-Command ($exe, $arguments)
-{
-  Write-Log "Executing command '$exe $($arguments -join ' ')'"
-  $output = .$exe $arguments
-
-  Write-CommandOutput $output
-  if (($LASTEXITCODE -ne $null) -and ($LASTEXITCODE -ne 0)) {
-    Write-Error "Command returned exit code $LASTEXITCODE. Aborting."
-    exit 1
-  }
-  Write-Log "done."
-}
-
-function Write-CommandOutput
-{
-  param (
-    [string] $output
-  )
-
-  if ($output -eq "") { return }
-
-  Write-Verbose ""
-  $output.Trim().Split("`n") |% { Write-Verbose "`t| $($_.Trim())" }
-  Write-Verbose ""
-}
-
-function Run-OctopusDeploy
-{
-  $exe = 'C:\Program Files\Octopus Deploy\Tentacle\Tentacle.exe'
-#"C:\Program Files\Octopus Deploy\Tentacle\Tentacle.exe" service --instance "Tentacle" --install --start
+ if(!(Test-Path $TentacleExe)) {
+	throw "File not found. Expected to find '$TentacleExe' to perform setup."
+  }  
+  
   Write-Log "Start Octopus Deploy Tentacle instance ..."
-  $args = @(
+  Execute-Command $TentacleExe @(
     'service',
     '--console',
     '--instance', 'Tentacle',
-    '--install',
     '--start'
   )
-  Execute-Command $exe $args
+
   "Run started." | Set-Content "c:\octopus-run.initstate"
 
   # try/finally is here to try and stop the server gracefully upon container stop
@@ -60,22 +25,20 @@ function Run-OctopusDeploy
      # sleep-loop indefinitely (until container stop)
     $lastCheck = (Get-Date).AddSeconds(-2)
     while ($true) {
-      Get-EventLog -LogName Application -Source "OctopusDeploy Tentacle*" -After $lastCheck | Select-Object TimeGenerated, EntryType, Message
+      Get-EventLog -LogName Application -Source "Octopus*" -After $lastCheck | Select-Object TimeGenerated, EntryType, Message
       $lastCheck = Get-Date
-       "$([DateTime]::Now.ToShortTimeString()) - OctopusDeploy service is '$((Get-Service "OctopusDeploy Tentacle").status)'."
+       "$([DateTime]::Now.ToShortTimeString()) - OctopusDeploy Tentacle service is '$((Get-Service "OctopusDeploy Tentacle").status)'."
        Start-Sleep -Seconds 60
     }
   }
   finally {
       Write-Log "Shutting down Octopus Deploy instance ..."
-      $args = @(
+      Execute-Command $TentacleExe @(
         'service',
         '--console',
         '--instance', 'Tentacle',
-        '--install',
         '--stop'
       )
-      Execute-Command $exe $args
   }
 
   Write-Log ""
@@ -84,10 +47,10 @@ function Run-OctopusDeploy
 try
 {
   Write-Log "==============================================="
-  Write-Log "Running Octopus Deploy"
+  Write-Log "Running Octopus Deploy Tentacle"
   Write-Log "==============================================="
 
-  Run-OctopusDeploy
+  Run-OctopusDeployTentacle
 
   Write-Log "Run successful."
   Write-Log ""
